requirements/requirements.txt
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
autopep8==1.6.0
binance-connector==1.14.0
cassandra-driver==3.25.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
constantly==15.1.0
cryptography==37.0.4
distlib==0.3.5
docopt==0.6.2
filelock==3.7.1
flake8==4.0.1
Flask==2.1.3
Flask-Cors==3.0.10
geomet==0.2.1.post1
hdfs==2.7.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.1
mccabe==0.6.1
oauthlib==3.2.0
platformdirs==2.5.2
py4j==0.10.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pyOpenSSL==22.0.0
pyspark==3.2.2
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
toml==0.10.2
tweepy==4.10.0
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.26.10
Werkzeug==2.2.0
zope.interface==5.4.0
tqdm
pandas
ntscraper